==> s02_IH_Jan18_GM_variant_filter_v2.Rmd <==
---
title: "Filter variants on consequence and rarity"
author: "EF"
date: "08 Jan 2018"
output: html_document
---
# start_section
  
```{r start_section}

# Time stamp
Sys.time()

# Clean up
rm(list=ls())
graphics.off()

# Set root working folder
library(knitr)
library(dplyr)
library(stringr)

opts_knit$set(root.dir = "/rds/project/erf33/rds-erf33-medgen/users/eleanor/2015-16/IHCAP/IH_Jan18_GM/")
setwd("/rds/project/erf33/rds-erf33-medgen/users/eleanor/2015-16/IHCAP/IH_Jan18_GM/")

```

#Load image from s01
```{r load image}

load("R_data/s01_IH_Jan18_GM_20180108.Rdata")

#Check dimensions
dim(gt.filt)
dim(vv.filt)
dim(exac.filt)
dim(kgen.filt)

sum(row.names(gt.filt) != row.names(vv.filt)) 
sum(row.names(gt.filt) != row.names(exac.filt))
sum(row.names(gt.filt) != row.names(kgen.filt))

#coerce gt and vv to dataframes
gt.df <- as.data.frame(gt.filt)
vv.df <- as.data.frame(vv.filt)

#Check row names have not changed
sum(row.names(gt.filt) != row.names(gt.df)) 
sum(row.names(vv.filt) != row.names(vv.df))

#Clean up
rm(vv.filt, gt.filt)
```

#Add exac and kgen frequencies to vv file
```{r exac kgen}

#Add kgen AFs
vv.df$kgen.AF <- kgen.filt$kgen.AF
vv.df$kgen.EUR_AF <- kgen.filt$kgen.EUR_AF

#Add exac AFs
vv.df$exac_non_TCGA.AF <- exac.filt$exac_non_TCGA.AF
vv.df$exac_non_TCGA_EUR.AF <- exac.filt$exac_non_TCGA.AC_NFE/exac.filt$exac_non_TCGA.AN_NFE

```


#Filter vv on consequence
```{r vv filter consequence}

#Number of variants before filtering on consequence
nrow(vv.df)

#Filter on consequence
high <- vv.df[vv.df$IMPACT == "HIGH",]
miss <- vv.df[vv.df$IMPACT == "MODERATE",]
del <- miss[(grepl("deleterious", miss$SIFT_call) & grepl("probably_damaging", miss$PolyPhen_call)) |
              (is.na(miss$SIFT_call) & grepl("probably_damaging", miss$PolyPhen_call)) |
              (grepl("deleterious", miss$SIFT_call) & is.na(miss$PolyPhen_call)) |
              grepl("inframe", miss$Consequence),]
high_del <- rbind(high, del)

#Number of variants after filtering on consequence
nrow(high_del)

#Clean up
rm(vv.df, miss, high, del)
```

#Filter on rarity
```{r vv filter kgen}
#Create vector of common variants in EUR_Kgen
kgen_com <- row.names(kgen.filt[kgen.filt$kgen.EUR_AF > 0.05,])

#Number of varaints to remove
length(kgen_com)

#Select variants not in this vector
vv <- high_del[!row.names(high_del) %in% kgen_com,]

#Number of variants remaining after consequence and rare filters
nrow(vv)

#Sort vv on rownames
vv <- vv[order(row.names(vv)),]

#Clean up
rm(high_del, kgen_com, kgen.filt, exac.filt)
```

#Select filtered variants in gt
```{r gt filter}
#Select variants in gt that passed filters
gt <- gt.df[row.names(gt.df) %in% row.names(vv),]

#Check rownames
dim(gt)
dim(vv)

sum(row.names(gt) != row.names(vv)) 

#Clean up
rm(gt.df)
```

#Save data
```{r save}
save.image("R_data/s02_IH_Jan18_GM_20180108.Rdata")
```
